requirements.txt
python-dateutil
Pillow
exifread